// src/app/Models/almacen/OrdenCarga/ordencarga.model.ts
export class OrdenCarga {
    IdOrdenCarga: number;
    Folio: number;
    FechaEnvio: Date;
    IdCliente: number;
    Cliente: string;
    IdPedido: number;
    Fletera: string;
    Caja: string;
    Sacos: string;
    Kg: string;
    Chofer: string;
    Origen: string;
    Destino: string;
    Observaciones: string;
    Estatus: string;
    FechaInicioCarga: Date;
    FechaFinalCarga: Date;
    FechaExpedicion: Date;
    IdUsuario: number;
    Usuario: string;
}